Changes return values of analysis functions to allow easier plotting
Two functions are changed here. The feature selection function now returns
all the RFECV models trained. This allows these trained models to be pickled
and later analyzed to compare the results of using different accuracy thresholds.
The learning params function is also modified to allow users to specify
an 'end frame'. This allows user to plot learning changes in the first N frames
of an experiment, instead of over the whole experiment by default.